requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.23
sentence-transformers>=2.2.2

faiss-cpu>=1.7.0
torch
pyampute
